requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
feedparser>=6.0.11
requests>=2.32.3
